No funca esa huevada :((((

+++ sia_library_back/src/controllers/group.controller.ts
@@ -7,6 +7,7 @@ import { Period } from "../models/period.model";
 import { connection } from "../connection/connection";
 import { Grade } from "../models/grades.model";
 import { Teacher } from "../models/teacher.model";
+import { QueryTypes } from "sequelize";
 
 // ? Obtener todos los grupos
 export const getGroups = async (req: Request, res: Response) => {
@@ -125,15 +126,32 @@ export const addStudentToGroup = async (req: Request, res: Response) => {
 
 //? Eliminar estudiantes de grupo
 export const deleteStudentFromGroup = async (req: Request, res: Response) => {
-  const { id: group_id } = req.params;
-  const { student_id } = req.body;
-
   try {
-    const query = `DELETE FROM grupo_estudiante WHERE id_grupo = ? AND id_estudiante = ?`;
-    await connection.query(query, { replacements: [group_id, student_id] });
-    res.status(200).json({ message: "Estudiante eliminado del grupo" });
+    const { id: group_id } = req.params;
+    const { student_id } = req.body;
+
+    console.log("Estudiante a eliminar:", student_id);
+    
+
+    // Verificar si el estudiante existe en el grupo antes de eliminarlo
+    const existingStudent = await connection.query(
+      `SELECT * FROM grupo_estudiante WHERE id_grupo = ? AND id_estudiante = ?`,
+      { replacements: [group_id, student_id], type: QueryTypes.SELECT }
+    );
+
+    if (existingStudent.length === 0) {
+      return res.status(404).json({ message: "Estudiante no encontrado en el grupo" });
+    }
+
+    // Eliminar el estudiante del grupo
+    await connection.query(
+      `DELETE FROM grupo_estudiante WHERE id_grupo = ? AND id_estudiante = ?`,
+      { replacements: [group_id, student_id], type: QueryTypes.DELETE }
+    );
+
+    return res.status(200).json({ message: "Estudiante eliminado del grupo" });
   } catch (error) {
-    console.log("Algo malio sal: ", error);
+    console.log("Algo salió mal:", error);
     res.status(500).json({ message: "Error al eliminar estudiante del grupo" });
   }
 };
